fixed header inclusion

+++ src/main.c
@@ -10,8 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "src/fillit.h"
-#include "libft/libft.h"
+#include "fillit.h"
 
 int	main(int argc, char **argv)
 {
@@ -22,16 +21,16 @@ int	main(int argc, char **argv)
 
 	tetri_nb = 0;
 	if (argc != 2)
-		return (write(1, "usage: ./fillit input_file\n", 28));
+		return ((int)write(1, "usage: ./fillit input_file\n", 28));
 	fd = open(argv[1], O_RDONLY);
 	if (fd < 0)
-		return (write(1, "error opening\n", 14));
+		return ((int)write(1, "error opening\n", 14));
 	chr_count = read(fd, buff, BUFF + 1);
 	buff[chr_count] = '\0';
 	if (ft_isvalid(buff, chr_count, tetri_nb) != 1)
-		return (write(1, "error\n", 6));
+		return ((int)write(1, "error\n", 6));
 	else
-		return (write(1, "ok\n", 3));
+		return ((int)write(1, "ok\n", 3));
 //	ft_parse(buff);
 	return (0);
 }
